libzycorec: Core library needed by Zyan libraries
The ZyCoreC library provides platform independent types and macros
used by for Zyan libraries. We use this library only to provide the
data type definitions for libzydis. Since we have a libc (including
nolibc) in Unikraft, the source files are not required to be
compiled in our use case. Because libzydis will be the only library
so far that needs these headers, we hide libzycorec from the menu.

Signed-off-by: Simon Kuenzer <simon.kuenzer@neclab.eu>
Reviewed-by: Felipe Huici <felipe.huici@neclab.eu>